Address PR review: split task/step files, fix metric, add integration tests
- Move task sampling logic to fme/ace/stepper/task.py; keep
  InfillPredictionStep in fme/core/step/infill_prediction.py
- Remove separate loss_task metric (task loss is already in total loss)
- Use Stepper.build_loss for both normal and task losses
- Add test_train.py integration tests for SingleModuleStep prediction-only
  and InfillPredictionStep mixed-task modes
