+++ fme/ace/__init__.py
@@ -92,6 +92,7 @@ from fme.ace.stepper.single_module import (
     StepSelector,
     TrainStepperConfig,
 )
+from fme.ace.stepper.task import TaskConfig, TaskSamplingConfig, TaskWeights
 from fme.ace.stepper.time_length_probabilities import (
     TimeLengthMilestone,
     TimeLengthProbabilities,
@@ -123,13 +124,8 @@ from fme.core.step import (
     SeparateRadiationStepConfig,
     SingleModuleStepConfig,
 )
+from fme.core.step.infill_prediction import InferenceSchemeConfig
 from fme.core.step.multi_call import MultiCallConfig
-from fme.core.step.task_step import (
-    InferenceSchemeConfig,
-    TaskConfig,
-    TaskSamplingConfig,
-    TaskWeights,
-)
 from fme.core.typing_ import Slice
 
 from . import step
