+++ fme/core/step/__init__.py
@@ -1,6 +1,6 @@
+from .infill_prediction import InfillPredictionStep, InfillPredictionStepConfig
 from .multi_call import MultiCallStep, MultiCallStepConfig
 from .radiation import SeparateRadiationStep, SeparateRadiationStepConfig
 from .secondary_module import SecondaryModuleStep, SecondaryModuleStepConfig
 from .single_module import SingleModuleStep, SingleModuleStepConfig
 from .step import StepABC, StepConfigABC, StepSelector
-from .task_step import InfillPredictionStep, InfillPredictionStepConfig
